fikset datatype
long -> integer for task
UUID -> integer for user

+++ src/main/java/no/bachelor26/Tasks/Exception/TaskNotFoundException.java
@@ -6,7 +6,7 @@ public class TaskNotFoundException extends RuntimeException {
         super("Oppgaven med navn: '" + taskName + "', finnes ikke.");
     }
 
-    public TaskNotFoundException(Long taskID){
+    public TaskNotFoundException(Integer taskID){
         super("Oppgaven med ID: '" + taskID + "', finnes ikke.");
     }
